Embed material DB builder to fix crafting dropdown
