+++ public/crafting.html
@@ -33,7 +33,6 @@
         grid-template-columns: repeat(2, minmax(0, 1fr));
       }
     </style>
-    <script type="module" src="database.js"></script>
     <script type="module" src="crafting_calculator.js"></script>
 </head>
   <body style="background-image: url('Art/Backgrounds/fall.png'); background-size: cover; background-attachment: fixed;">
