+++ public/siege.html
@@ -27,7 +27,6 @@
     dt { font-weight: bold; color: #64e9a9; margin-top: 1rem;}
     dd { margin-left: 1.5rem; color: #cbd5e1; }
   </style>
-  <script type="module" src="database.js"></script>
   <script type="module" src="siege_calculator.js"></script>
 </head>
 <body style="background-image: url('Art/Backgrounds/volcano.png'); background-size: cover; background-attachment: fixed;">
